Prepare techs without area
